(clean) rename files
