+++ README.md
@@ -1,12 +1,3 @@
-## To Do
+# GO Ladder for Discord (GOLD) Website
 
-
-[x] Add Goban on player pages
-[x] Back button
-[ ] Display the 10 last played games
-[ ] Display who won/loe on each game
-[ ] Display who is White/Black
-[ ] Display the Handicap
-[ ] Conserver le clic du choix des joueurs classés
-[ ] outil de recherche de joueur sur la topbar
-[ ] courbe au passage de la souris sur le profil de joueur
+This website is the frontend of the GOLD project
