+++ src/Pages/PlayerList.css
